refactor(bi): remove legacy script generation APIs
