+++ wow-bi/build.gradle.kts
@@ -1,6 +1,7 @@
 description = "Wow BI Sync Script Generator"
 
 dependencies {
+    api(project(":wow-api"))
     implementation(project(":wow-core"))
     testImplementation(project(":example-domain"))
     testImplementation(project(":example-transfer-domain"))
