--- wow-bi/src/main/kotlin/me/ahoo/wow/bi/ScriptTemplateEngine.kt
@@ -1,53 +1,0 @@
-/*
- * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *      http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package me.ahoo.wow.bi
-
-import me.ahoo.wow.api.Wow
-import me.ahoo.wow.api.modeling.NamedAggregate
-import me.ahoo.wow.bi.renderer.ClickHouseScriptRenderer
-
-object ScriptTemplateEngine {
-    const val DEFAULT_KAFKA_BOOTSTRAP_SERVERS = "localhost:9093"
-    const val DEFAULT_TOPIC_PREFIX = Wow.WOW_PREFIX
-
-    fun renderGlobal(): String = ClickHouseScriptRenderer().renderGlobal()
-
-    fun renderClear(namedAggregate: NamedAggregate, expansionTables: List<String>): String =
-        ClickHouseScriptRenderer().renderClear(namedAggregate, expansionTables)
-
-    fun renderCommand(
-        namedAggregate: NamedAggregate,
-        kafkaBootstrapServers: String = DEFAULT_KAFKA_BOOTSTRAP_SERVERS,
-        topicPrefix: String = DEFAULT_TOPIC_PREFIX,
-    ): String = legacyRenderer(kafkaBootstrapServers, topicPrefix).renderCommand(namedAggregate)
-
-    fun renderStateEvent(
-        namedAggregate: NamedAggregate,
-        kafkaBootstrapServers: String = DEFAULT_KAFKA_BOOTSTRAP_SERVERS,
-        topicPrefix: String = DEFAULT_TOPIC_PREFIX,
-    ): String = legacyRenderer(kafkaBootstrapServers, topicPrefix).renderStateEvent(namedAggregate)
-
-    fun renderStateLast(namedAggregate: NamedAggregate): String =
-        ClickHouseScriptRenderer().renderStateLast(namedAggregate)
-
-    private fun legacyRenderer(
-        kafkaBootstrapServers: String,
-        topicPrefix: String,
-    ): ClickHouseScriptRenderer = ClickHouseScriptRenderer(
-        BiScriptOptions(
-            kafkaBootstrapServers = kafkaBootstrapServers,
-            topicPrefix = topicPrefix,
-        )
-    )
-}
