--- wow-bi/src/main/kotlin/me/ahoo/wow/bi/expansion/TableNaming.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *      http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package me.ahoo.wow.bi.expansion
-
-import me.ahoo.wow.api.Wow
-import me.ahoo.wow.api.modeling.NamedAggregate
-import me.ahoo.wow.bi.BiScriptOptions
-
-object TableNaming {
-    fun NamedAggregate.toTopicName(prefix: String = Wow.WOW_PREFIX, suffix: String): String {
-        return BiTableNaming(BiScriptOptions(topicPrefix = prefix)).toTopicName(this, suffix)
-    }
-
-    fun NamedAggregate.toDistributedTableName(suffix: String): String {
-        return BiTableNaming().toDistributedTableName(this, suffix)
-    }
-}
